When editing the file, only send in the new attributes.
This is especially important for the `content_copy` attribute, since
what we get from DRF is a string, but apparently it doesn't like that
when we're sending it back as a string. So now, in this case since we
only just edit the format, the server will only send the format
attribute in a PATCH call.

But the root problem of DRF not accepting the content_copy field it just
erved us is still present, and is the real solution to this. But I'm
putting it off for now.