remove get_user_edit_channels nginx caching
this causes issues with the user not seeing the new channel if they refresh immediately. From what we've seen in the field, this can lead to users creating multiple copies of a channel.

This is exacerbated when we have less frontend pods running, since users are more likely to hit the same studio frontend pod instance.