Fix action for updating assessment items
This action failed when creating new assessment items
because newly it's not possible to save an assessment item
to store without `assessment_id` => with our current
IndexedDB resource layer that's responsible for generating
IDs assessment item needs to be commited to IndexedDB before
it can be saved to store. This also implicates that it's not
possible to use sanitization as proposed without more updates
of related client IDs logic => removing for now.