Remove ancestors watcher
Due to reactivity, this made `loadAncestors`
to infinitely loop and that broke studio tree
navigation on channel edit page.
Instead watch `topicId` and load full ancestors
path on its value change or on the first component
load. This commit also removes a condition that caused
ancestors to not be loaded in case that a topic was
already in a path because that might not be sufficient
when navigating using URLs and similar.