Assign copyright_holder w/ @input & ensure mastery criteria is saved
Selecting from the autocomplete suggestions doesn't trigger a native event so we need to listen on the normal input event as well.

Also - mastery criteria was not saving. I had a feeling it was to do with the promises not being resolved before continuing and it looks like I was right.