Move MoveModal out of a parent v-if block
MoveModal open/close logic was conflicting
with a condition based on selected nodes
length which was causing the modal to be
closed prematurely without properly calling
logic that is supposed to run on its close.

Fixes #2866.
Fixes issue when selecting a node that was
previously moved unexpectedly reopens the move
modal.
Fixes issue when a snackbar isn't displayed.