Fix outside to inside reactivity
Make the component's reactivity work also
when receiving updated value from its mother
component.
In a mostly async environment, we can't be sure
that the first value received is the correct,
up-to-date value, and not the previous outdated value.
It can happen that the up-to-date value that should
be rendered arrives a couple of milliseconds later,
and without reactivity also working from outside to inside,
the outdated value would be displayed in such cases.

Fixes #2888