Remove IconButton wrapper
IconButton.vue was originally a temporary wrapper, oferring some
extra logic such as RTL support. However, KDS takes care of RTL,
so it's not needed. One more thing IconButton did was automatically
setting aria label - if we're certain this is desired in all cases,
it can be introduced as a KDS feature.